+++ Integrations/IntSight/IntSight.py
@@ -50,7 +50,16 @@ SEVERITY_LEVEL = {
 }
 
 
-def req(method, path, json_data=None, params=None):
+def is_json(myjson):
+    demisto.info('walla')
+    try:
+        json_object = json.loads(myjson)
+    except ValueError as e:
+        return False
+    return True
+
+
+def req(method, path, json_data=None, params=None, json_response=False):
     """
     Send the request to IntSights and return the JSON response
     """
@@ -59,6 +68,8 @@ def req(method, path, json_data=None, params=None):
     if r.status_code < 200 or r.status_code > 299:
         if not (r.text == 'SeverityNotChanged' or r.text == 'TagExist' or r.text == 'IocBlocklistStatusNotChanged'):
             return_error('Error in API call to IntSights service %s - [%d] %s' % (path, r.status_code, r.text))
+    if json_response and not is_json(r.text):
+        return_error('Error in API call to IntSights service - please check you URL address')
     return r
 
 
@@ -264,7 +275,7 @@ def get_alert_by_id_helper(alert_id):
     """
     Helper for getting details by ID
     """
-    r = req('GET', 'public/v1/data/alerts/get-complete-alert/' + alert_id).json()
+    r = req('GET', 'public/v1/data/alerts/get-complete-alert/' + alert_id, json_response=True).json()
     return alert_to_readable(r, True), alert_to_readable(r, False)
 
 
@@ -321,7 +332,7 @@ def get_alert_activity():
     Retrieves the alert activity by alert-id
     """
     alert_id = demisto.getArg('alert-id')
-    r = req('GET', 'public/v1/data/alerts/activity-log/' + alert_id).json()
+    r = req('GET', 'public/v1/data/alerts/activity-log/' + alert_id, json_response=True).json()
 
     human_readables = []
     alert = {'ID': alert_id, 'Activities': []}
@@ -380,7 +391,7 @@ def change_severity():
 
 
 def get_assignee_id(assignee_email):
-    r = req('GET', 'public/v1/account/users-details')
+    r = req('GET', 'public/v1/account/users-details', json_response=True)
     r = r.json()
     for user in r:
         if assignee_email == user.get('Email', ''):
@@ -487,7 +498,7 @@ def send_mail():
 
 
 def get_tag_id(alert_id, tag_name):
-    res = req('GET', 'public/v1/data/alerts/get-complete-alert/' + alert_id)
+    res = req('GET', 'public/v1/data/alerts/get-complete-alert/' + alert_id, json_response=True)
     res = res.json()
 
     details = res.get('Details', {})
@@ -735,7 +746,7 @@ def get_iocs():
     """
     Gets all IOCs with the given filters
     """
-    r = req('GET', 'public/v1/iocs/complete-iocs-list', params=handle_filters()).json()
+    r = req('GET', 'public/v1/iocs/complete-iocs-list', params=handle_filters(), json_response=True).json()
     domains = []
     ip_infos = []
     url_infos = []
@@ -838,7 +849,7 @@ def update_ioc_blocklist_status():
 
 def get_ioc_blocklist_status():
     alert_id = demisto.getArg('alert-id')
-    r = req('GET', 'public/v1/data/alerts/blocklist-status/' + alert_id).json()
+    r = req('GET', 'public/v1/data/alerts/blocklist-status/' + alert_id, json_response=True).json()
     demisto.results({
         'Type': entryTypes['note'],
         'EntryContext': {
@@ -851,7 +862,7 @@ def get_ioc_blocklist_status():
 
 def get_mssp_sub_accounts():
     account_id = demisto.getParam('credentials')['identifier']
-    accounts = req('GET', 'public/v1/mssp/customers').json()
+    accounts = req('GET', 'public/v1/mssp/customers', json_response=True).json()
     if not accounts:
         return_error("intsights-mssp-get-sub-accounts failed to return data.")
 
@@ -871,7 +882,7 @@ def get_mssp_sub_accounts():
     for i, account in enumerate(account_ids):
         # Call account
         HEADERS['Account-Id'] = account
-        account_ua = req('GET', 'public/v1/account/used-assets').json()
+        account_ua = req('GET', 'public/v1/account/used-assets', json_response=True).json()
 
         if not account_ua:
             continue