+++ Integrations/AWS-Lambda/AWS-Lambda.py
@@ -60,7 +60,7 @@ def aws_session(service='lambda', region=None, roleArn=None, roleSessionName=Non
     if kwargs and AWS_ACCESS_KEY_ID is None:
 
         if AWS_ACCESS_KEY_ID is None:
-            sts_client = boto3.client('sts')
+            sts_client = boto3.client('sts', config=config, verify=VERIFY_CERTIFICATE)
             sts_response = sts_client.assume_role(**kwargs)
             if region is not None:
                 client = boto3.client(