+++ Integrations/IntSight/CHANGELOG.md
@@ -1,1 +1,2 @@
-set default encoding to utf-8
+## [Unreleased]
+- Indicators issue fixed - were not extracted correctly in `!intsights-get-iocs`