+++ Integrations/IntSight/CHANGELOG.md
@@ -1,1 +1,5 @@
-set default encoding to utf-8
+## [Unreleased]
+
+
+## [19.9.1] - 2019-09-18
+Improved the error message in cases where the URL address is incorrect.